html/semantics/embedded-content/media-elements/audio_duration_values.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Audio Test: audio_duration_values</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/2014/CR-html5-20140429/embedded-content-0.html#htmlmediaelement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<audio id="a">
  Your browser does not support media elements.
</audio>
<audio id="m" src="/media/sound_5.mp4">
  Your browser does not support media elements.
</audio>
<div id="log"></div>
<script>
test( function() {
  var media = document.getElementById("a");
  assert_equals(media.duration, NaN, "the value of media.duration");
}, "Check if the audio.duration returns NaN if the duration isn't available");

async_test(function (t) {
  var media = document.getElementById("m");
  media.addEventListener("loadedmetadata", t.step_func(function () {
    assert_equals(Math.round(media.duration), 5, "The value of media.duration");
    t.done();
  }));
}, "Check if the audio.duration returns the length of the media resource");
</script>
